exceptions

+++ lalchowk_desktop/messages.cs
@@ -41,9 +41,8 @@ namespace Veiled_Kashmir_Admin_Panel
         }
             catch (Exception ex)
             {
-                var message = ex.ToString();
-                string[] split = message.Split(new string[] { " at " }, StringSplitOptions.None);
-                MessageBox.Show("Something happened, please try again.\n\n" + split[0], "Error!");
+
+                MessageBox.Show("Something happened, please try again.\n\n" + ex.Message.ToString(), "Error!");
             }
         }
 
@@ -120,28 +119,31 @@ namespace Veiled_Kashmir_Admin_Panel
 
         private void delbtn_Click(object sender, EventArgs e)
         {
-            DialogResult dgr = MessageBox.Show("Do you want to delete this message?", "Confirm!", MessageBoxButtons.YesNo);
-            if (dgr == DialogResult.Yes)
-            {
-                try
+            if (messagesdataview.SelectedCells.ToString().Contains("")){
+                MessageBox.Show("Nothing to delete");
+            }
+            else {
+                DialogResult dgr = MessageBox.Show("Do you want to delete this message?", "Confirm!", MessageBoxButtons.YesNo);
+                if (dgr == DialogResult.Yes)
                 {
-                    string cmd = "delete from messages where messageid='" + midlbl.Text + "'";
-                    obj.nonQuery(cmd);
-                    obj.closeConnection();
-                    MessageBox.Show("Message deleted.");
-                    Cursor = Cursors.WaitCursor;
-                    readmsgs();
-                    messagesdataview.DataSource = bsource;
-                    Cursor = Cursors.Arrow;
-                }
-                catch (Exception ex)
-                {
-                    var message = ex.ToString();
-                    string[] split = message.Split(new string[] { " at " }, StringSplitOptions.None);
-                    MessageBox.Show("Something happened, please try again.\n\n" + split[0], "Error!");
+                    try
+                    {
+                        string cmd = "delete from messages where messageid='" + midlbl.Text + "'";
+                        obj.nonQuery(cmd);
+                        obj.closeConnection();
+                        MessageBox.Show("Message deleted.");
+                        Cursor = Cursors.WaitCursor;
+                        readmsgs();
+                        messagesdataview.DataSource = bsource;
+                        Cursor = Cursors.Arrow;
+                    }
+                    catch (Exception ex)
+                    {
+
+                        MessageBox.Show("Something happened, please try again.\n\n" + ex.Message.ToString(), "Error!");
+                    }
                 }
             }
-            
         }
     }
 }
